Moved the Register class from luigi.task to its own luigi.task_register
This was mainly done to make documentation more clear (so that Register doesn't clutter luigi.task)
